requirements.txt
# Art CRM Requirements
# Python 3.12+

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Configuration
python-dotenv==1.0.1    # Environment variable management

# Phase 4: Import Script
pandas==2.2.0           # Excel file reading and data manipulation
openpyxl==3.1.2         # Excel file format support for pandas
rapidfuzz==3.6.1        # Fuzzy string matching for venue names

# Phase 2+:
click==8.1.7          # CLI framework
requests==2.32.4      # HTTP requests for DeepSeek and web APIs (< 2.32.4 had CVE-2024-35195 and .netrc leak)
anthropic==0.39.0     # Claude API client

# Phase 6-Alpha: Lead Generation
googlemaps==4.10.0    # Google Maps Places API client
beautifulsoup4==4.12.3  # HTML parsing for web scraping
tqdm==4.66.4          # Progress bars (4.66.1 had CVE-2024-34062)

# Testing
pytest==7.4.3         # Testing framework
pytest-cov==4.1.0     # Coverage reporting
flake8==7.0.0         # Linting (pre-commit hook + CI)

# Future dependencies:
# fastapi==0.109.0      # Web API framework
# uvicorn==0.27.0       # ASGI server for FastAPI
